Update types and fix some warnings

+++ Debug/DPrint.c
@@ -19,8 +19,10 @@ void ResetDPrintYPos()
 void DPrint(const char *str, const uint color)
 {
 #ifdef ENABLE_DEBUG_PRINT
-	FontDrawString((Vector2){12, dprintYPosition + 2}, str, 16, 0xFF000000, smallFont);
-	dprintYPosition += (int)FontDrawString((Vector2){10, dprintYPosition}, str, 16, color, smallFont).y - dprintYPosition + 8;
+	FontDrawString((Vector2){12, (float)dprintYPosition + 2}, str, 16, 0xFF000000, smallFont);
+	dprintYPosition += (int)FontDrawString((Vector2){10, (float)dprintYPosition}, str, 16, color, smallFont).y -
+					   dprintYPosition +
+					   8;
 #endif
 }
